chorbi 2.0 to 1.0

+++ Proyects/Acme-Chorbi/src/main/webapp/views/message/list.jsp
@@ -29,7 +29,7 @@
 
 	
 	<display:column>
-		<a href="chirp/chorbi/delete.do?messageId=${row.id}">
+		<a href="chirp/chorbi/delete.do?messageId=${row.id}" onclick="return confirm('<spring:message code="message.confirm.delete" />') ">
 			<spring:message code="message.delete" />
 		</a>
 	</display:column>
@@ -38,7 +38,7 @@
 	
 	<spring:message code="message.title" var="titleHeader" />
 	<display:column title="${titleHeader}" sortable="true" >
-		${mask:mask(row.subject) }
+		${mask:mask(row.subject) } 
 	</display:column>
 
 	<spring:message code="message.moment" var="momentHeader" />
@@ -51,9 +51,11 @@
 	
 	<spring:message code="message.attachment" var="attachmentHeader"/>
 	<display:column title="${attachmentHeader}">
+	<ul>
 	<jstl:forEach items="${row.attachments}" var="thisAttachment">
-		<a href="${thisAttachment.link}" target="_blank">${thisAttachment.link}</a>
+		<li><a href="${thisAttachment}" target="_blank">${thisAttachment}</a>
 	</jstl:forEach>
+	</ul>
 	</display:column>
